requirement.txt
cffi==1.14.4
click==7.1.2
cryptography==3.2.1
Flask==1.1.2
Flask-MySQL==1.5.2
Flask-MySQLdb==0.2.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysqlclient==2.0.3
pycparser==2.20
PyJWT==2.0.0
six==1.15.0
Werkzeug==1.0.1
